Remove thin wrapper cogs and add conversation memory
- Move 5 task files from tasks/ to cogs/ replacing thin wrappers:
  daily_digest, daily_haiku, yahoo_fantasy_weekly, vibecheck_weekly, daily_hero_dm
- Add _get_conversation_turns() to GeminiCog for multi-turn context
- Update call_llm() to include conversation history from message archive
- Fix vibecheck_weekly import path after move
